+++ src/RequestHooks.php
@@ -4,6 +4,7 @@ namespace PwJsonApi;
 
 /**
  * Request hooks
+ * @todo Prevent cross-service injection
  */
 class RequestHooks extends Hooks
 {