+++ src/ServiceList.php
@@ -2,6 +2,8 @@
 
 namespace PwJsonApi;
 
+use ProcessWire\WireException;
+
 /**
  * Represents a list of services
  */
@@ -46,15 +48,27 @@ class ServiceList
    * @template TService of Service
    * @param TService $service
    * @param (callable(TService): void)|null $setup
+   * @todo Prevent cross-service injection when API is running
    */
-  public function add(Service $service, callable|null $setup = null): static
-  {
-    $this->items[] = $service;
+  public function add(
+    Service $service,
+    callable|null $setup = null,
+    Api|Service|null $context = null,
+  ): static {
+    // Prevent cross-service injection
+    // if ($context instanceof Service && $context->_isInitialized === true) {
+    //   throw new WireException(
+    //     "Cannot add service '{$service->name}' as a child service for '{$context->name}'. Service '{$context->name}' is already initialized.",
+    //   );
+    // }
+
+    $service->_prepare();
 
     if (is_callable($setup)) {
       call_user_func($setup, $service);
     }
 
+    $this->items[] = $service;
     return $this;
   }
 }