+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### New features
 
+- New `init()` method for Services
 - New `Request` object is exposed in arguments of request handlers and hooks
 - New plugin interface
 - CSRF protection plugin